Reimplemented A.I. Added support for mnist database.

+++ src/main/java/cz/stv/neurondemofx/ImageDrawable.java
@@ -21,5 +21,5 @@ package cz.stv.neurondemofx;
  */
 public interface ImageDrawable
 {
-  public void draw(byte[][] matrix);
+  public void draw(short[][] matrix);
 }
